just move location centering to TableMap component
